- Moved all routes to `/api/v1`. - Created new healthcheck at `/health` - Updated server dependencies.
