+++ api-tests/capacities/capacities - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/capacities
+  url: http://localhost:5000/api/v1/capacities
   body: json
   auth: none
 }
