+++ api-tests/capacities/capacities - delete.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 delete {
-  url: http://localhost:5000/capacities/1
+  url: http://localhost:5000/api/v1/capacities/1
   body: json
   auth: none
 }
