+++ api-tests/capacities/capacities - findAll (paginate).bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/capacities?pageNumber=1&pageSize=5&sortField=createdAt&sortOrder=desc
+  url: http://localhost:5000/api/v1/capacities?pageNumber=1&pageSize=5&sortField=createdAt&sortOrder=desc
   body: none
   auth: none
 }
@@ -15,7 +15,7 @@ query {
   pageSize: 5
   sortField: createdAt
   sortOrder: desc
-  ~filterOperator: LIKE
+  ~filterOperator: like
   ~filterField: name
   ~filterValue: 1
 }
