+++ api-tests/capacities/capacities - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/capacities
+  url: http://localhost:5000/api/v1/capacities
   body: none
   auth: none
 }
