+++ api-tests/capacities/capacities - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/capacities/1
+  url: http://localhost:5000/api/v1/capacities/1
   body: none
   auth: none
 }
