+++ api-tests/capacities/capacities - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/capacities/1
+  url: http://localhost:5000/api/v1/capacities/1
   body: json
   auth: none
 }
