+++ api-tests/drives/drives - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/drives
+  url: http://localhost:5000/api/v1/drives
   body: json
   auth: none
 }
