+++ api-tests/drives/drives - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/drives
+  url: http://localhost:5000/api/v1/drives
   body: none
   auth: none
 }
