+++ api-tests/drives/drives - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/drives/1
+  url: http://localhost:5000/api/v1/drives/1
   body: none
   auth: none
 }
