+++ api-tests/drives/drives - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/drives/1
+  url: http://localhost:5000/api/v1/drives/1
   body: json
   auth: none
 }
