+++ api-tests/interfaces/interfaces - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/interfaces
+  url: http://localhost:5000/api/v1/interfaces
   body: json
   auth: none
 }
