+++ api-tests/interfaces/interfaces - delete.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 delete {
-  url: http://localhost:5000/interfaces/1
+  url: http://localhost:5000/api/v1/interfaces/1
   body: json
   auth: none
 }
