+++ api-tests/interfaces/interfaces - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/interfaces
+  url: http://localhost:5000/api/v1/interfaces
   body: none
   auth: none
 }
