+++ api-tests/interfaces/interfaces - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/interfaces/1
+  url: http://localhost:5000/api/v1/interfaces/1
   body: none
   auth: none
 }
