+++ api-tests/interfaces/interfaces - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/interfaces/3
+  url: http://localhost:5000/api/v1/interfaces/3
   body: json
   auth: none
 }
