+++ api-tests/manufacturers/manufacturers - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/manufacturers
+  url: http://localhost:5000/api/v1/manufacturers
   body: json
   auth: none
 }
