+++ api-tests/manufacturers/manufacturers - delete.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 delete {
-  url: http://localhost:5000/manufacturers/4
+  url: http://localhost:5000/api/v1/manufacturers/4
   body: json
   auth: none
 }
