+++ api-tests/manufacturers/manufacturers - findAll (pagination).bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/manufacturers?pageNumber=1&pageSize=2&sortField=name&sortOrder=desc&filterField=name&filterValue=Seagate&filterOperator=like
+  url: http://localhost:5000/api/v1/manufacturers?pageNumber=1&pageSize=2&sortField=name&sortOrder=desc&filterField=name&filterValue=Seagate&filterOperator=like
   body: none
   auth: none
 }
