+++ api-tests/manufacturers/manufacturers - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/manufacturers
+  url: http://localhost:5000/api/v1/manufacturers
   body: none
   auth: none
 }
