+++ api-tests/manufacturers/manufacturers - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/manufacturers/1
+  url: http://localhost:5000/api/v1/manufacturers/1
   body: none
   auth: none
 }
