+++ api-tests/manufacturers/manufacturers - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/manufacturers/1
+  url: http://localhost:5000/api/v1/manufacturers/1
   body: json
   auth: none
 }
