+++ api-tests/models/models - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/models
+  url: http://localhost:5000/api/v1/models
   body: json
   auth: none
 }
