+++ api-tests/models/models - delete.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 delete {
-  url: http://localhost:5000/models/1
+  url: http://localhost:5000/api/v1/models/1
   body: json
   auth: none
 }
