+++ api-tests/models/models - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/models
+  url: http://localhost:5000/api/v1/models
   body: none
   auth: none
 }
