+++ api-tests/models/models - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/models/1
+  url: http://localhost:5000/api/v1/models/1
   body: none
   auth: none
 }
