+++ api-tests/models/models - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/models/1
+  url: http://localhost:5000/api/v1/models/1
   body: json
   auth: none
 }
