+++ api-tests/retailers/retailers - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/retailers
+  url: http://localhost:5000/api/v1/retailers
   body: json
   auth: none
 }
