+++ api-tests/retailers/retailers - delete.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 delete {
-  url: http://localhost:5000/retailers/4
+  url: http://localhost:5000/api/v1/retailers/4
   body: json
   auth: none
 }
