+++ api-tests/retailers/retailers - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/retailers
+  url: http://localhost:5000/api/v1/retailers
   body: none
   auth: none
 }
