+++ api-tests/retailers/retailers - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/retailers/1
+  url: http://localhost:5000/api/v1/retailers/1
   body: none
   auth: none
 }
