+++ api-tests/retailers/retailers - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/retailers/1
+  url: http://localhost:5000/api/v1/retailers/1
   body: json
   auth: none
 }
