+++ api-tests/storageTypes/storageTypes - create.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 post {
-  url: http://localhost:5000/storageTypes
+  url: http://localhost:5000/api/v1/storageTypes
   body: json
   auth: none
 }
