+++ api-tests/storageTypes/storageTypes - delete.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 delete {
-  url: http://localhost:5000/storageTypes/4
+  url: http://localhost:5000/api/v1/storageTypes/4
   body: json
   auth: none
 }
