+++ api-tests/storageTypes/storageTypes - findAll (ids).bru
@@ -5,12 +5,12 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/storageTypes?id=1&id=2
+  url: http://localhost:5000/api/v1/storageTypes?id=1
   body: none
   auth: none
 }
 
 query {
   id: 1
-  id: 2
+  ~id: 2
 }
