+++ api-tests/storageTypes/storageTypes - findAll (paginate).bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/storageTypes?pageNumber=1&pageSize=2&sortField=createdAt,name&sortOrder=desc&filterField=id&filterValue=2&filterOperator=eq
+  url: http://localhost:5000/api/v1/storageTypes?pageNumber=1&pageSize=2&sortField=createdAt,name&sortOrder=desc&filterField=id&filterValue=2&filterOperator=eq
   body: none
   auth: none
 }
