+++ api-tests/storageTypes/storageTypes - findAll.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/storageTypes
+  url: http://localhost:5000/api/v1/storageTypes
   body: none
   auth: none
 }
