+++ api-tests/storageTypes/storageTypes - findOne.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://localhost:5000/storageTypes/1
+  url: http://localhost:5000/api/v1/storageTypes/1
   body: none
   auth: none
 }
