+++ api-tests/storageTypes/storageTypes - update.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 put {
-  url: http://localhost:5000/storageTypes/1
+  url: http://localhost:5000/api/v1/storageTypes/1
   body: json
   auth: none
 }
