+++ client/src/providers/data-provider/index.ts
@@ -3,7 +3,7 @@
 // import dataProviderSimpleRest from "@refinedev/simple-rest";
 
 // const API_URL = "https://api.fake-rest.refine.dev";
-const API_URL = "http://localhost:5000";
+const API_URL = "http://localhost:5000/api/v1/";
 
 // export const dataProvider = dataProviderSimpleRest(API_URL);
